+++ src/app/providers/index.ts
@@ -1,1 +1,3 @@
 export * from './feedback.provider';
+export * from './list-files.providers';
+export * from './list-modules.providers';